+++ mod/makerelease.sh
@@ -9,7 +9,10 @@ modname=`cat info.json|jq -r .name`
 modver=`cat info.json|jq -r .version`
 
 # Create git tag for this version
-git tag "mod_$modver"
+#git tag "mod_$modver"
 
 # Prepare zip for Factorio native use and mod portal
 git archive --prefix "${modname}_$modver/" -o "${modname}_$modver.zip" HEAD
+
+cp "${modname}_$modver.zip" "../modpackage/${modname}.zip"
+cp "info.json" "../modpackage/info.json"